Split anim into different files in graphic dir
